+++ web/src/app/app.module.ts
@@ -15,6 +15,7 @@ import { AngularFireFunctionsModule, FunctionsRegionToken } from '@angular/fire/
 
 // Dashboard hub modules
 import { AppRoutingModule } from './app-routing.module';
+import { CoreModule } from './core/core.module';
 
 // Dashboard hub components
 import { AppComponent } from './app.component';
@@ -37,7 +38,8 @@ export function tokenGetter(): string | null {
         DeviceDetectorModule.forRoot(),
         HttpClientModule,
         RouterModule,
-        AppRoutingModule
+        AppRoutingModule,
+        CoreModule
     ],
     providers: [
         { provide: FunctionsRegionToken, useValue: 'us-central1' }
