+++ web/src/app/core/guards/authentication.guard.ts
