+++ web/src/app/dialog/list/dialog-list.component.ts
@@ -1,6 +1,6 @@
 import { Component, Inject } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material';
-import { RepositoryService } from '../../../services/repository.service';
+import { RepositoryService } from '../../core/services/repository.service';
 import { Subscription } from 'rxjs';
 import { ProjectModel, RepositoriesModel } from '../../../models/index.model';
