+++ web/src/app/projects/create/create.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
 import { MatSnackBar } from '@angular/material';
-import { ProjectService } from '../../../services/project.service';
+import { ProjectService } from '../../core/services/project.service';
 import { catchError } from 'rxjs/operators';
 import { ProjectModel } from '../../../models/index.model';
 import { Subscription } from 'rxjs';
