+++ web/src/app/projects/repository/repository.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { RepositoryModel } from '../../../models/index.model';
 import { Subscription } from 'rxjs';
-import { RepositoryService } from '../../../services/repository.service';
+import { RepositoryService } from '../../core/services/repository.service';
 
 @Component({
     selector: 'dashboard-repository',
